requirements.txt
# IPO Investment Dashboard - Python 의존성 패키지
# 설치: pip install -r requirements.txt

# HTTP 요청
requests>=2.31.0

# 데이터 처리
pandas>=2.1.0

# 환경변수 관리
python-dotenv>=1.0.0

# AI 예측 (predict.py에서 사용)
google-genai>=1.0.0
